Add application level pepper config; startup opts
In doing some prep work for TOTP secondary authentication, some
discussion regarding protecting TOTP secrets at rest came up.  Often
times, the TOTP secrets are simply stored in plain text while first
factor credentials get a better treatment.  This may be because with
storing passwords, all we need to store is irreversible hash of the
salted password; we get the actual password from the user and we will
know the salt and therefore can calculate the hash and compare with what
is stored.  With TOTP we need to know the actual secret prior to
the user providing anything of note in calculating what TOTP code to
expect at any given time.

The suggestions were to use a pepper value, not stored in the database,
as part of the secret generated and assigned to the user.  At least this
means that you cannot understand the secret unless you also have gotten
the pepper value from server memory or a file or similar stored
(hopefully) away from the database.

Add this new parameter to msbms_start_options.sample.toml.  Looks like
the development version of msbms_start_options.toml has also got
changes to commit... not sure it that dev file is there intentionally or
not.  Will commit the change for now, but may remove the file later.