Add common table conf_enums
This new table is to replace the wide variety of enum_* tables found in
the app.  Over time I've waffled on this question of single big enum
table or smaller easier to reference and specialize approach.  Given the
current room temperature, and the fact that dealing with a single cached
table in the application software will be far simpler, that a single
table is the way to go with some specialized triggers for validation and
check constraint support.

Also add to global and instance database definitions since both will
need this sort of structure.