Fix incorrect name for trigger name
trig_b_iu_syst_identities_check_uniqueness was named without the syst
part in a couple places.  While there are some circumstances where we
will drop the table kind prefix in our naming, the trigger and trigger
function naming are not amongst them.

Fix it where it was found to be inconsistent.