Change code standard; add new collation type
The test case for this is the conf_enum_values table.  When adding the
new collation for the internal_name and display_name columns it became
clear that the column definition formatting standard was going to either
fail, become cumbersome due to excessive length, or cause strange breaks
that were non-congruous with other lines not requiring the collation
specification.

As a test, rely more on indented new lines to separate column definition
elements.  The upsides are that its actually easier this way to see
where columns start than that the previous (or typical) standard, the
line lengths are even more sane, the standard itself is easier to
understand and implement compared to the previous standard, the standard
is more resilient to new elements and one-off feature requirements than
the old standard, and the different column options are more individually
identifiable than under the previous standard.  The downside is chiefly
related to the increase of vertical space taken.  So for example the
conf_enum_values table under the previous standard easily fit on a
single screen, but under the new standard it doesn't.

While I might change back to a more common layout later, right now I
think the upsides of the new standard make for a clear improvement for
table DDL.