Refactor description field to allow for syst/user
The current pattern is to allow in-record documentation fields which may
have system set descriptions to allow for user overrides of those texts
without losing the system version.

This commit adds this pattern to two places where it wasn't already
implemented: syst_feature_types & syst_modules.

Changing this to the new standard is something that is happening when
the old pattern is encountered and is not systematic.  So expect other
commits like this to slowly evolve the database to the new standard over
time.