Refactor seed_syst_feature_types; instance/common
This is WIP commit.  Update the records to be added to 1) not be wrong
which some of the previously committed values were; 2) include the newer
naming and extra feature groups.  Now that there's a better idea about
where this fits, we can be a bit more consistent in how we pre-populate
syst_feature_types.

This work is both done in the common and instance database variants of
the seed data.