Refactor feature mapping
The previous structure wasn't going to result in a good feature for
creating feature maps and tree structures.  Too much was hard coded and
as defined it wasn't going to produce a desirable result; or what
constitutes desirable changed enough from when first written that
clearly the design was simply to brittle to accommodate change.