Add system/maintenance/displayable flags
This adds flag columns allowing the system to distinguish between system
defined records and user defined records, or to selectively grant the
user permission the ability to manage in full.

Also add a displayable flag to the syst_feature_map table.  There are
reasons to protect system defined entries here, but we still want to
afford users the ability to define their own hierarchies.  By allowing
users to set system feature map records to not be displayable, we can
keep the records, but not force them into user displays.