Add function initialize_enum
Right now the thinking is that getting most of the seed data which
defines the system expected enumerations written will be easier/safer to
do using nested JSON definitions which can then be processed by DO block
scripts or functions.

This function does the record creation for the conf_enums tables based
on a JSON representation which sticks pretty close to the corresponding
record definitions.  The biggest difference is that the one-to-many
relationships are represented as via nested arrays of objects.

The conf_enums tables seem to lend themselves well to a more formally
defined functional process than, say, the syst_feature_map group of
tables.