Refactor global db enums to conf_enums configs
This commit refactors the global database individual enum table design
to now use the conf_enums tables.  This commit is hand tested by
creating the starting migration for global and loading it into a
database.  The testing did not confirm that all triggers and functions
are correct, though a number of them were exercised during the test.