Change conf to syst and remove some old conf files
Distinguishing the difference between what was a "conf" table and a
"syst" table was just making too fine a distinction to reasonable or
reliable.  conf can be seen as a specialized subset of syst and where it
might be considered user data, it still usually performs a significant
system operations functionality.  So convert all conf to be syst base.

Along the way there are a couple old conf tables which should go away as
they are to be replaced with a more generic configuration system in the
common database definitions.