Fix enum value check trigger naming collision
Updated the naming standard for triggers calling
trig_a_iu_enum_value_check.  The new standard includes the name of the
enumeration to check since we were getting trigger naming collisions
when the table had more than one enumeration.

All tables with existing references to the trigger were renamed whether
or not the suffered the collision.