Add custom Ecto Interval data type
This is needed to support, right now, the msbms_syst_settings module
configuration point for intervals.  Ecto/Postgrex doesn't immediately
expose certain PostgreSQL recognized data types and so it has to be done
on a custom type basis.

We want to decouple the application from the the library specifics, so
we basically mirror the Postgrex type in a custom type;  this should
leave the basic implementation details hidden to the rest of the
application.