Refactor database source tree
While starting on msbms_syst_settings and figuring out that eventually
that development will require a working database with the settings in
place... and concluding that this will be a recurring theme as more and
more modules are worked on... it became clear that the database source
code structure was geared too much on just getting the global and
instance app databases going.

So... restructure the source to align more completely with the module
breakout in the Elixir code.  Over all this will be better since it
aligns database code with the module that manages it and also make
module/module dependencies clearer.

Along with the refactor, make the naming more consistent with standards
and organization elsewhere in the overall application.  In some cases
rename just to get alphabetical sorting (like renaming common to all and
using specially chosen prefixes that are both sortable and have
meaning.)