Fix incorrect dir ref in msbms_syst_datastore
The migrations code looks to the priv directory to find the DDL for the
migrations schema/table, but the path was hard coded.  This works for
running the component's tests, but isn't sufficient when the component
is used as a library in other components (given it's a library, we can
expect this).

The code looking for the priv directory must resolve that at runtime via
:code.priv_dir/1.  We can then just self-reference our application and
get its priv directory where ever it happens to be.