Refactor msbms_syst_settings test DB build scripts
The utility scripts which tear down the DB after testing and the
database schema creation DDL were targeted at the datastore test scripts
or the general application use case.

Here, we don't want to build or test the API level concerns that will be
in place once the general application is built, that's a concern for the
general app, and we obviously need to tear down the database we build
for settings testing, not datastore testing.  So commit those changes
which customize the test run for the settings module.