Refactor msbms_syst_datastore transaction fn name
Adding the other Ecto.Repo Queryable related callbacks to Datastore, it
became apparent that the other callbacks where named to make clear in
the implementation file that they came from Ecto but that transaction
was inconsistently renamed (db_transaction).  So make it the same as the
others now that we have many more than we started with.