Add wrapper Ecto funcs a& API delegates; Datastore
This adds most of the "Queryable callbacks" from Ecto.Repo to both
Datastore and the public API defined by MsbmsSystDatastore.

The reason for this is that 1) we need to ensure that the correct repo
is being used to query the database so our wrappers add the database
connection reference as the first parameter; and 2) we want to abstract
out the data access library from the rest of the application.  We're
assuming that (for the most part) the Ecto API here is good enough and
familiar enough for other developers to use for our purposes.  Should
that change, we'd keep the API and hide the logic in the
implementation... or at least that's the present plan.

Note that the work here is not complete.  We'll need to add typespecs 
and documentation, but for now, we can settle for just functional.