Refactor location of generic schema macro
We have a common schema __using__ macro which sets up options common to
all our database structures.  Previously this was defined in
msbms_syst_settings as that was the first module to define an Ecto
Schema struct, but really the common schema macro belongs in Datastore
since the settings module will be far from the last module to need it.