Refactor datastore db_types location
Previously this was held under the impl directory because it does
contain logic but it's unclear that was going to be a good home.  The
reality is that other modules are going to rely on those data types
(structs) so, they aren't just behind the scenes.

For now, go ahead an promote db_types to be part of the first class API,
including the logic since it's simple.  In other cases of public structs
with more extensive logic, it may make sense to make the struct public,
as we do here, but then call/delegate to a more extensive logic module
in impl.