Refactor Datastore Repo reference requirements
Many msbms_syst_datastore functions would take a Repo reference of some
sort in order to explicitly call put_dynamic_repo and make sure the
right Repo was used in whatever action was being taken.  This really
wasn't necessary though.

As I've learned more about the process dictionary, upon which the
dynamic repo mechanism works, it became clear that really in most cases
the database would be called in the context of certain long-lived
processes which would already be specific to a given Datastore context;
at least as far as tenancy is concerned, though there may still be a
need to switch datastore contexts based on the caller's user (for
example API calls vs. UI calls).  Anyway, constantly setting the Repo
pid was probably just burning cycles and and probably just moving where
the datastore context needed to be retained.

So this commit updates Datastore to not require it nearly so often and
just rely on the process dictionary put_dynamic_repo/get_dynamic_repo
mechanism (OK, I wrap it, but only just).