Remove dependence on msbms_syst_features
In msbms_syst_settings, we previously had a dependence on the features
module since there is a join table referencing both module's tables.

However, this architecture didn't really make sense.  This made a one-
way mandatory dependency to features in the settings module, and would
be one of many such dependencies.  Allowing feature mapping to be fully
optional would reduce the coupling so that's what we've done.

The dependency to feature mapping is removed entirely from settings and
then implemented by msbms_syst_features since knows all of different
kinds of features it can support and how it should support them.  This
puts the support in a common module (msbms_syst_features) and they can
check for the supported feature's existence at compile for optional
support compilation.