Finish logic in msbms_syst_settings (preliminary)
This update brings a fair number of changes to already committed files
as well as fills out much of the logic of the module.

  - Change config_* field names to be setting_*.

  - Add changeset related logic for creation and updating of settings to
    the SystSettings struct.

  - Complete the public API module (MsbmsSystSettings), document the
    functions, and add typespecs.

  - Define relevant module specific types in MsbmsSystSettings.Types

  - Add the internal business logic module
    MsbmsSystSettings.Impl.Settings

  - Add a GenServer based runtime management module at
    MsbmsSystSettings.Runtime.Server

Do note that at this time, the tests have yet to be written.  Once some
reasonable testing is in place, we can complete this module's feature
branch.