Refactor msbms_syst_settings service start funcs
When starting to write tests and needing to put the GenServer backing
the settings service under supervision, it became apparent that the
start-up function call in the API and return values were just not
sufficient.  In fact, they were just different than what one would
expect from such a service.  Also, trying to treat already started as
not an error can cause issues in this scenario as well as if the service
is already started, we'll try to supervise it a second time since we'll
not know that it should already be supervised.

Use the typical start_link semantics and return the GenServer.start_link
return values.  Let the caller sort out things like already started...
which is a cleaner implementation anyway.