Fix settings view trigger to return old on delete
Turns out that if you don't return old from the instead of trigger on
delete, DELETE FROM x RETURNING * returns no row.  Not surprising and
surprising at the same time.

We need that to verify that we actually delete a row in the
msbms_syst_settings module (and eventually elsewhere) so be sure we're
returning the old row.