Fix delete_settings func in msbms_syst_settings
Turns out that the Ecto.Repo.delete callback is particularly finicky in
what it will accept to delete.  It wants you to at least select the
record and then go to a delete, and even then I couldn't get it to work
(didn't try hard, because it's wrong from the database sense).

This commit implements the common Ecto pattern of creating a query and
and then calling Ecto.Repo.delete_all.