Fix refresh function in msbms_syst_settings
refresh_from_database/1 was a delegate directly to the Settings module,
but that won't work because the ets table that backs the settings
service is protected and owned by the GenServer process, so any calls
that write to the ets table must be there (or have special
accommodations made).  It really should have gone there anyway as we'd
want this process to serialize write calls there anyway.