Fix/refactor sys_settings.display_name changing
First, the trigger for updating the display_name column of syst_settings
didn't include the display_name column at all which means it would have
been impossible to change the display_name value via the application in
any situation, syst_defined true or not.  This is fixed.

Second, decided (for now) to allow display name changes on syst_defined
true settings.  This may be reversed again later, but seems like right
now it's better to allow (and test it) rather than to deny it just yet.
Ultimately, it will be a business rules/usability decision on whether or
not the decrease in supportability of system defined settings outweighs
the user configurability increase... not our problem yet.