Fix typespecs in msbms_syst_datastore.ex
Bad start_link typespec was causing dialyzer some problems so fix that
to allow tests to pass.

Also add a typespec for "all" delegate.  This is probably not needed,
but dialyzer was complaining about that as well... turned out not to be
a typespec problem but something about stale plt files or dependencies
needing to be refreshed.  A proper clean-out is what really fixed
things.