Fix possibly bad returns from syst_settings API
The instead of triggers do data changing updates that should, mostly,
function as regular table changes.  However, the trigger variables 'new'
and 'old' reflect the view data values and not the underlying records;
so in this way they don't really behave the same as 'new' and 'old' in
table triggers.  Once you're aware of this fact, it makes perfect sense,
but may not be the default mindset (it wasn't mine).

So, for these views to work in a table-like way for INSERT/UPDATE/DELETE
operations, we need to explicitly perform the table operation and then
populate 'new'/'old' with either a RETURNING clause in the instead of
trigger table operation, or via some other way of correctly populating
those trigger variables prior to return.

Here, we add the RETURNING clauses to the table operations for the
syst_settings API view.  Note that we use `RETURNING * INTO new/old`
rather than explicitly naming columns.   This is done since the view is
just an access control on the underlying table: if the table changes we
want the view to break loudly since the view will also have to be
updated and * is the easiest way to get close to that ideal.