Refactor syst_settings view update validity check
The prior check would fail on validations for columns that would never
be updated by this view.  That wasn't necessarily wrong; trying to use
the API to make invalid changes arguably should raise an error so that
the application doesn't carry on thinking that its changes were valid.
However, the problem is that such errors can't be readily detected
across all actions.  For example in an INSERT, we may see
`new.syst_defined` as FALSE from the view, but even though syst_defined
is not maintainable via the view, that value isn't wrong either: should
we fail on principle or only fail if the value is invalid?  Add to this
that the defaulting may have happened at lower application abstractions
that may be needed for various use cases and it's much clear what the
right answer is.

For now, we'll adopt a "postel's law" attitude to the problem and be
fairly permissive about what we'll receive from an API caller.  We will
still raise if a fundamental assumption about a change is likely to be
significant.  For example, changing the `internal_name` of a system
defined record should be fatal since the after-change assumption of the
record having a different identity may well be expected to have
consequences for the caller, but changes to fields like syst_defined
itself is less likely to.

To be clear, it's not clear that this is the right approach and we may
well want to strengthen the checks, but for now this approach seems to
have the correct level of pragmatism.