Refactor database table/trigger directory struct.
Prior to this commit, there was a separation of table and trigger
database object definitions into specific directories.  While working
with the API views, which include the INSTEAD OF trigger definitions in
the same directory as the view definitions, it's become apparent that
having table specific trigger files kept alongside the table definition
is a more clear/intuitive way to keep these files than the previous
method.

What we're really saying is that triggers are more an aspect of table
definition than just a functional construct.  This changes if the
trigger is general purpose (applying to many tables), and we've not
changed the definition of those triggers other than to bump them out of
a "multi-table" sub-directory.  This change in thinking about table
specific triggers has also meant that they should be in the same schema
as the table to which they're attached.

This commit updates this everywhere and tests insofar as is feasible
(they global & instance build plans successfully build migrations).