Add global/instance migration build/deploy tests
These tests are only there to ensure that the starting database
migrations can 1) be built; 2) be installed.  These databases are then
immediately dropped and not used for functional testing of the
application logic.  This will be useful as the "Release 0" version of
the software is being developed.

The functional testing database will be migrated in test_helper.exs and
will use pre-built/versioned migration files which would typically
accompany a release.