Add/Refactor SystSettings changeset validations
Rename the maybe_validate_user_description functions to just be
validate_user_description.  No need for the "maybe" here.

Add a private validation function for the internal_name column.  When a
setting is user defined, they should be able to change the internal_name
(so long as naming business rules are met).  So the new validation
function tries to make sure that only user defined settings are changed
and that, if valid, those changes follow the other rules.