Refactor SystSettings changeset handling
Based on work done in the msbms_syst_enums module, it's become clear
that the changeset handling should more closely resemble that in the
documented examples, specifically that there should be one changeset
function and that should adapt to different scenarios.  While I think
that our original approach of defining create and update changeset
functions is cleaner, when you start working with Ecto associations that
is much less the case.  So we won't fight it and just go with the
expected pattern.