Add syst_enum_functional_types check for allow new
Enumerations may be created not requiring that functional types are
defined; this also means that an enumeration item records will have
no functional type since none are defined.  However adding a functional
type to an existing enumeration with existing enumeration items means
that those enumeration items would have to be assigned functional types
in order to maintain validity.   While it might be possible with a
sufficiently sophisticated set of deferred constraints and other
processing code, allowing the creation of a functional type requirement
on an existing enumeration item records would be complex and error
prone.  Best just to not allow the creation of such a requirement if
enumeration items exist.

The trigger in this commit checks for this condition and raises if a new
functional type requirement is going to be forced on existing
enumeration item records.

Also update the enum section of various interested build plans, along
with some minor reordering/reformatting.