Fix incorrect autogenerate in datastore schema
The Elixir documentation doesn't do a great job of making clear across
contexts what autogenerate means.  Perhaps to someone just working in
application code and not both application and database code it's more
clear, but for me it's always confusing.

With some multi-record processing now written, it's clear that the
autogenerate value was not being set to depend on the database defaults
but rather was improperly generating the id in the application.