Refactor syst_enum_functional_types trigger name
The prior name of new_allowed_check was too specific and a bit unclear
as to what we were doing.  This is a simple data validation trigger so
let's make that clear; I'm not sure how much benefit there is in
decoupling different validations into separate functions so for now
let's assume there isn't any and so long as we're validating what is
passed rather than making assumptions/changing things we keep the
validations in one function.  If there's a compelling reason to split it
later we can rethink our policy then.