Refactor enum func type validation trig name/role
This is effectively a reversal of a prior commit which suggested that
validations could be put into a single trigger function.  Bad idea.  As
soon as that was done, the first prospect of needing to do it was going
to turn that validation trigger into an imperative hell-hole.  This is
in part due to the conflicts between not wanting to duplicate sometimes
complex logic and the differences of INSERT/UPDATE/DELETE operations and
information availability... throw in IF and CASE statements not being
deterministic in terms of condition evaluation (so that code like
TG_OP = 'INSERT' OR OLD.test_value != NEW.test_value  won't be handled
reliably).

Keeping each validation in it's own trigger function just keeps all of
this cleaner and easier to read.  We do end up moving a little logic
into the table definitions where we set triggers; this is the case when
we conditionally execute a trigger function.  But that's solely focused
on whether to run the trigger or not so I don't think that logic leak is
too bad at all.

Along with this a better definition (or confirmation of a nearly
forgotten definition) of the trigger name prefixes mean:

  * a00_ - a99_ : These are early validations designed to stop later,
    possibly more expensive processing later in the trigger sequences.
    We shouldn't be changing or setting data in this class of function,
    though we might look up data in other tables.

  * b00_ - b99_ : Data manipulation functions.  These are the functions
    we want updating or setting values that are appropriate for the
    limits of database maintained business logic (data integrity/
    consistency focused).

  * c00_ - c99_ : Late validation functions.  These are the functions
    which validate data after the data manipulation phase.  In reality
    most data validation should probably be here since the "b" class
    triggers can take a starting invalid record and make it valid.

  * z00_ - z99_ : Utility/Auxiliary processing functions.  This class
    of trigger processes records/data/etc that aren't directly related
    to the business data in the record.  For example, the diagnostic
    columns such as last updated or updated by or functions that record
    data changes to an audit table are the kinds of functions which
    should be in the "z" class.  These functions should not make
    business validation judgements or take transformative actions on
    business data.  These functions are really operational in nature
    and work on records which, from a business data perspective, are in
    their final form.

This should be documented elsewhere, but for now include here so the
thinking isn't lost; other commits will implement this as well as
disparities with the thinking above are found.