Refactor syst_enum_items table triggers
This is an update to naming and timing around various triggers.

First, I can't see any real value to making the functional type
validation a constraint trigger. In practice the data will already exist
or it should be created in order (there's no reason to not create it in
order) so the ability to defer the check just shouldn't happen in
practice; also the application will be the one creating possibly
violating records and it doesn't support deferred constraints. So make
this check a before check since we might possibly avoid some extra data
interactions by doing this in the before timing.

Next, swap the ordering of the checks based on ideas around trigger
classes ("a", "b", "c", "z") described elsewhere.

Fix a minor formatting issue in
trig_a_iu_syst_enum_items_maintain_sort_order.  "NULL" is treated as a
keyword and so should be all caps.