Fix invalid updates in trig_i_u_syst_enum_items
The update trigger was allowing a re-assignment of parent enumeration
and this isn't really valid, certainly not through the API view.  The
correct action would be to delete it and recreate it on the desired
enumeration.

Also, the trigger was not testing for updates to internal_name and
functional_type_id changes for system defined enumeration items.  While
some internal_name and functional_type_id changes are valid for updates,
they aren't for syst_defined = true records.