Refactor db table locations across existing code
While working with files, it's become clear that the <schema>/tables
structure is not helpful.  We have data-centric schemas so the tables
designation is redundant, and since we've moved trigger functions to be
close to the table definition, it's not even accurate anymore.

We remove the unnecessary level across all existing definitions where it
exists or is referenced.