Refactor example enum test data syst_defined
We should make this false to reduce the number of enumerations we need
to create to support examples.  I don't think we lose anything moving
off the existing setting and we gain the ability to more freely change
data across the record types; this would be helpful in examples for
data changing functions.