Downgrade ecto dependencies to 3.7.2 from 3.8.x
It looks like ecto_sql version 3.8.0 has a bug related to dynamic repos.

I've reported the bug and may try my hand at fixing it, but in the
meantime, let's downgrade.  Originally I upgraded because of something
I though had a chance of being a bug in Ecto, but that turned out to be
a bug in my code; so we really don't need the 3.8. version.

Clean out all the mix.lock files as well, just in case.