Remove feature mapping from initialize_enum func
The msbms_syst_enums initialization function wanted to handle the
feature mapping feature, but that has since been made a higher level
dependency which means we shouldn't reference it at all in
msbms_syst_enums.  Also, at that time the syst_enums table had a
foreign key reference to that column.

Remove the expectation and capability to deal with feature mapping in
the initialize_enum function.

Also remove the references to `feature_internal_name` in both the global
and instance seed data.