Refactor sponsor number in test build plans
Sponsor 820 is "MS" in base 36; one of our reserved sponsor numbers.
That is our regular everyday deployment values for sponsor.  For testing
we should make it clear with something different.  821 which translates
to base 36 "MT" seems to be appropriate: "Muse Testing".

Update the testing build plans to build for sponsor 821 now that this is
decided.