Refactor syst_instances fields
Previously we had hard coded columns for things like dbserver name and
app/api context connection counts.  If we're taking the decoupling of
modules seriously, building in very specific requirements of the
application that will use this component library will probably defeat
our purpose.  More than that, it's not even really clear at this point
what the correct contexts will be for the application and it's changed
a time or two already.

Refactor the application specific columns out and replace them with a
single jsonb column for instance options.  This should allow us to have
discrete fields to manage instances that are not tightly coupled to the
client application as well as giving us some flexibility to change our
expectations and requirements in the application a bit more dynamically.

Of course, this pushes the burden to the application for data
consistency, but I think in this instance it's OK: we're dealing with
variable application runtime dynamics and not business data so I think
this will be more apparent to any later developers.