Refactor msbms_syst_instance_mgr enums location
The global database had seed data initialization for the owner_states,
instance_types, and instance_states enums.  Those really now should go
in the msbms_syst_instance_mgr module as this is from where the
functional type definitions may be (should be) codified for any users of
the component library; for example, record purge owners functionality
would naturally fit into the functions of this module and so the owner
states, via functional types, which indicate that an owner record may be
purged must be defined in the enum.