Add owners test data; rearrange some ordering
In msbms_syst_instance_mgr test_data, add support for the syst_owners
table.

Also, reorder the syst_applications insert values.   While it may or may
not make a difference, we want to try and disrupt "database order" so
that sorting in the application as -maybe- a little more chance to
fail.