Add optimistic_locking call to all Ecto Schema def
I currently don't see a reason with the existing Ecto Schema definitions
to avoid only updating if the record version is known to be the same as
the starting point.  If a reason pops up, we can choose to remove it at
that time;  I expect that having different changeset functions may be a
real thing at some point and then we can finesse things... but we'll
assume for now.