Refactor owner seed data to assume a default enum
Prior to this commit, there was no enumeration-wide default defined for
the `owner_states` enumeration.  This makes `owner_states_sysdef_active`
the enumeration default for the starting seed data.

Note that default handling through changes is something we still have to
work out and make consistent; not an issue for this commit, but a
related consideration.