Add owner state functional types
In MsbmsSystInstanceMgr add a type to make clear what the valid owner
state functional types are.  This can be helpful in finding the default
enumeration item for a given functional type, for example.  Also, anyone
making use of MsbmsSystInstanceMgr can see what functional types for
owner state are suppose to be supported.