Refactor service ident in MsbmsSystSettings
Prior to this commit, the MsbmsSystSettings functionality required the
caller to pass the identity of the service on pretty much every call.
This is because we expect multiple instances of the MsbmsSystSettings
services to be running at any one time (one for each running instance/
tenant and the global subsystem).  This is pretty clunky though and
starts to poke through other modules we might wrapping the some of these
caching services (e.g. MsbmsSystSettings, MsbmsSystEnums).

Why not handle this in much the same way that Ecto does with dynamic
repos?  Ecto uses the process dictionary to cache the Repo to use for
all calls to the Repo functions.  That way you don't need to complicate
the calling business logic with this runtime concern.  While Ecto allows
you to change the cached current Repo, the reality is that it's probably
not going to change or change much in most scenarios.  I'm feeling that
this will be true for our service modules as well.

  * Add MsbmsSystSettings.Runtime.ProcessUtils for the functions which
    manage the process dictionary.

  * Update all the functions, removing the service_name parameters and
    replace the parameter's references with calls to
    ProcessUtils.get_settings_service.

  * Update documentation (and clean up documentation, too) to update
    examples with removed parameters.

  * Update test module tests for new parameterization requirements.