Refactor default changeset handing for enum items
The changeset function was requiring :enum_default and
:functional_type_default values.  In most cases setting these explicitly
has no value and simply complicates what you're trying to accomplish.

Allow the changeset function to test if the values exist, have been
provided via changes, and if needed (like an insert which doesn't
include them) default them.  In both cases, the default will be false.

I'm thinking there's probably a validation trigger here, but am not
fully settled.  It requires a decision on whether a default is always
required or not.  I expect I will require the default always and allow
setting an enum item to be default (for either enum/functional type)
will automatically make the pre-existing default record not default.
The special case will be can you delete the last record since it will be
the default as the last record (the answer is yes, but need to know well
the correct process for forcing that issue).  Deferring that to another
day for now.

Note that get_field will return a nil if the key exists in the data or
change map and is also set nil.  Since I think we're defaulting that on
new struct create... which feeds insert changesets, we will always get
that value from the data struct and never return the get_field default.
So get the result of get_field and OR it with false; this will get the
value included and set and it should never be null from the database so
our action here should be safe.