Refactor service ident in MsbmsSystEnums
This is similar to the earlier refactor of MsbmsSystSettings.  Rather
than depend on the caller to always tell us the service name so we can
look it up and call it as appropriate, rely on the process dictionary to
keep that runtime concern for us.  This simplifies the API and is
consistent with other service oriented components such as Ecto.Repo.

Note that, and this is true for MsbmsSystSettings as well, the correct
process arrangement might not be quite correct.  For example, Ecto.Repo
allows us to set the dynamic repo using put_dynamic_repo in `setup_all`
in our testing and that works for the entire test suite.  But for enums
and settings, I have to call `setup` before each test to set the service
name.  I'm not sure the reduced scope in which the service gets set is
even an issue, but having to call put_enums_service/1 possibly more
often seems like it could be a opportunity for error.  Different problem
for a different day at this stage; we'll see how the current approach
works in practice.